refactor!: move Dock settings to General settings
