+++ src/renderer/Core/Settings/Pages/General/General.tsx
@@ -1,6 +1,7 @@
 import { SettingGroup } from "@Core/Settings/SettingGroup";
 import { SettingGroupList } from "../../SettingGroupList";
 import { Autostart } from "./Autostart";
+import { DockSettings } from "./DockSettings";
 import { HotKey } from "./HotKey";
 import { Language } from "./Language";
 import { SearchHistory } from "./SearchHistory";
@@ -13,6 +14,7 @@ export const General = () => {
                 <Language />
                 <HotKey />
                 <Autostart />
+                <DockSettings />
             </SettingGroup>
             <SettingGroup title="Search History">
                 <SearchHistory />
